requirements.txt
numpy~=2.0.2
seaborn~=0.13.2
pillow~=10.4.0
matplotlib~=3.9.2
opencv-python~=4.10.0.84
scikit-image~=0.24.0
pandas~=2.2.2
improcrf~=0.1.0
scipy~=1.13.1